requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
discord==2.3.2
discord.py==2.5.2
docopt==0.6.2
frozenlist==1.5.0
idna==3.10
multidict==6.2.0
propcache==0.3.1
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.3.0
yarg==0.1.10
yarl==1.19.0
